Create tree.sh
Let’s work on a basic script that prints a directory structure in a tree format, which is useful for viewing nested subdirectories. The script takes advantage of the powerful commands that the 
[Linux](https://www.codecademy.com/resources/docs/open-source/linux) shell provides for file and folder operations.

We will be using conditional statements as well as the following commands:

- `echo`
- `cd`
- `pwd`
- `ls`
- `grep`
- `sed`
- `wc`

This script will take in an optional string argument: the parent directory that should be the starting point for the tree. If this is not provided, we will use the current working directory. We will print a message if the parent directory has no subfolders to print. If we find there are subfolders to print, we’ll use a combination of the commands `ls`, `grep`, and `sed` to print the directory structure tree.

We will be working on the script named `tree.sh`, which is already opened in the editor.