Update tree.sh
Let’s continue to work on the script that prints the directory structure tree. At the end of the last exercise, our script was able to print all nested subdirectories and files under a parent directory. In this exercise, we will add to the script. We’ll use the grep command to filter this output to only include subdirectories, not files, and format it to better match a tree-like structure using sed.

The left side of the image below shows what our script’s output looks like now. The right side is what it will look like after adding the grep and sed commands.A screenshot showing 2 side-by-side terminal outputs of the tree.sh script. On the left is before the grep and sed commands were added. On the right is the output after those commands were added.As you can see, the directory structure is a lot easier to follow after we will use the sed command to:

Remove colons
Use dashes (-) to differentiate between directory and sub-directory
Indent to the right by three spaces
Replace the first dash with a vertical bar (|)
Let’s make the output look nice!